requirements.txt
flask
flasgger
gunicorn
pyyaml